core(ConsolePaging): Move ConsolePaging in a dedicated module
